pyproject.toml
[project]
name = "shopify-mcp-server"
version = "1.0.0"
description = "Shopify MCP Server - Python implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.10.6",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "graphql-core>=3.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
shopify-mcp-server = "main:main"
